Add 2D view mode to visualizer
Add a new `ViewMode::View2D` enumeration to the view mode
options in the `ConfigurationVisualizer` class. This enhances
the visualizer by allowing users to switch between perspective,
orthogonal, and 2D views, improving flexibility for different
viewing needs.

Introduce the functionality in `FrmMain` to sync view mode
changes between components and updated UI elements in
`partmainvisualizer.ui` to include an icon for the new 2D view.
This change is made to enhance visualization options for users,
catering to workflows requiring a more planar visualization.